+++ src/api_server/src/request/mod.rs
@@ -12,6 +12,7 @@ pub mod metrics;
 pub mod mmds;
 pub mod net;
 pub mod snapshot;
+pub mod version;
 pub mod vsock;
 pub use micro_http::{
     Body, HttpServer, Method, Request, RequestError, Response, StatusCode, Version,