+++ CHANGELOG.md
@@ -8,6 +8,9 @@
   size for PUT and PATCH requests.
 - Limit MMDS data store size to `--http_api_max_payload_size`.
 - Cleanup all environment variables in Jailer.
+- Added metrics for accesses to deprecated HTTP and command line API endpoints.
+- Added permanent HTTP endpoint for `GET` on `/version` for getting the
+  Firecracker version.
 
 ### Fixed